Refactored & re-organized file structure from unit tests
